+ double large prime cutoffs changed to uint64s, simplifying the code and providing 	a small speedup for larger jobs + better polynomial root update and bucket sieving assembler code, providing a 	small speedup + fixed a couple more bugs in prime counting and printing (thanks again to Alex 	Balfour and his Calendar Magic beta testers!) + added some extra info to the -v printout at the end of DLP siqs factorizations + removed some obsolete code and cleaned up some compiler warnings

+++ trunk/factor/tune.c
@@ -280,7 +280,7 @@ void make_job_file(char *sname, uint32 *startq, uint32 *qrange, char *inputstr, 
 	//more accurate test
 	*qrange = 5000;
 
-	printf("nfs: commencing job file construction for c%d\n",strlen(inputstr));
+	printf("nfs: commencing job file construction for c%d\n",(int)strlen(inputstr));
 
 	fprintf(out, "n: %s\n",inputstr);
 	switch (inputnum)
